+++ main/src/main/java/com/main/comunicacion/privadas/DTOs/TipoIncidenciaPrivateDTO.java
@@ -6,10 +6,10 @@ import lombok.Data;
 
 @Data
 public class TipoIncidenciaPrivateDTO {
-    private Integer id;
+    private long id;
     private String nombre;
 
-    public TipoIncidenciaPrivateDTO(Integer id, String nombre) {
+    public TipoIncidenciaPrivateDTO(long id, String nombre) {
         this.id = id;
         this.nombre = nombre;
     }
